Add files via upload
Description: 
Design and implementation of a mobile App
The case study is used as an example in order to practice agile project management on the
Development. 
Description of the App:
The case study idea was to develop a mobile app (iOS, Android, cross-platform) 
The set of functionalities our app contain:
* User authentication, including registration, login, password
* A user interaction via a user interface (UX design) resulting a clickable prototype
Project Management:
The main goal was to plan, design and implement a (clickable) prototype of the app, to evaluate different
Options and Technologies for implementation and to manage the project in agile sprints.
All the Project Management works have done in Atlassian Confluence (all development documents and artefacts, e.g. user stories, FMEA), SharePoint, and Atlassian Trello.